Initial work on CLIENT-104 and CLIENT-107
For CLIENT-107 we still need to:
  - Use real version numbers in real releases
  - Some kind of striped-down launcher
  - Icons
  -.desktop files
  - documentation files

Currently it creates a deb file with a DMDirc.jar in it and nothing else, and only really for nightlies.

For real stuff we need to work out a way to convert our versioning scheme to debian-eqsue versioning, and some kind of auto-magic changelog generating.

Change-Id: Ie3af41828b6d2f96958527040e7ce877fe00a72d